Home-menu
Base redux config and moved styles from Home screen component.
