+++ src/containers/Home/redux-container/index.js
@@ -1,0 +1,8 @@
+import {bindActionCreators} from 'redux';
+import {connect} from 'react-redux';
+import Home from '../layout';
+
+export default connect(
+  ({}) => ({}),
+  dispatch => bindActionCreators({}, dispatch),
+)(Home);
