+++ src/containers/Home/styles/HomeStyles.js
@@ -1,0 +1,19 @@
+import {StyleSheet} from 'react-native';
+const style = StyleSheet.create({
+    reportList: {
+        backgroundColor: '#C4C4C4',
+        flexGrow:0,
+        height:"70%"
+    },
+    homeContainer: {
+        flex:1,
+        padding: 10,
+        alignContent:'flex-start'
+    },
+    homeTitle: {
+        fontSize: 20,
+        paddingVertical: 10
+    }
+})
+
+export default style
